csed342/assn4/soft_kmeans.py
# 20230262 Yeonung Kim

import math
import matplotlib.pyplot as plt
import numpy as np
import os
from utils import plot_2d_soft, read_data, load_centroids, write_centroids_tofile, set_seed

from kmeans import euclidean_distance

set_seed(42)

############################################################
# Problem 2: soft k-means clustering
############################################################

############################################################
# Problem 2a: calculate responsibility

def get_responsibility(data_point, centroids, beta):
    """Calculate the responsibiliy of each cluster for a single data point.
    You should use the euclidean_distance function (that you previously implemented).
    You can use the math.exp() function to calculate the responsibility.

    Arguments:
        data_point: a list of floats representing a data point
        centroids: a dictionary representing the centroids where the keys are
                   strings (centroid names) and the values are lists of
                   centroid locations
        beta: hyper-parameter

    Returns: a dictionary whose keys are the the centroids' key names and
             value is a float as the responsibility of the cluster for the data point.
    """
    # BEGIN_YOUR_ANSWER
    responsibility = {}
    denom = 0
    for key, centroid in centroids.items():
        denom += math.exp(-beta * euclidean_distance(data_point, centroid))
    for key, centroid in centroids.items():
        responsibility[key] = math.exp(-beta * euclidean_distance(data_point, centroid)) / denom
    return responsibility
    # END_YOUR_ANSWER

############################################################
# Problem 2b: update assignment

def update_soft_assignment(data, centroids, beta):
    """Find the responsibility of each cluster for all data points.
    You should use the get_responsibility function (that you previously implemented).

    Arguments:
        data: a list of lists representing all data points
        centroids: a dictionary representing the centroids where the keys are
                   strings (centroid names) and the values are lists of
                   centroid locations

    Returns: a dictionary whose keys are the data points of type 'tuple'
             and values are the dictionary returned by get_responsibility function.
             (In python, 'list' cannot be the 'key' of 'dict')
             
    """
    # BEGIN_YOUR_ANSWER
    responsibilities = {}
    for data_point in data:
        responsibilities[tuple(data_point)] = get_responsibility(data_point, centroids, beta)
    return responsibilities
    # END_YOUR_ANSWER
            
############################################################
# Problem 2c: update centroids

def update_centroids(soft_assignment_dict):
    """Update centroid locations with the responsibility of the cluster for each point
    as a weight. You can numpy methods for simple array computations. But the values of 
    the result dictionary must be of type 'list'.

    Arguments:
        assignment_dict: the dictionary returned by update_soft_assignment function

    Returns: A new dictionary representing the updated centroids
    """
    # BEGIN_YOUR_ANSWER
    updated = {}
    denoms = {}
    for key, value in soft_assignment_dict.items():
        for centroid, responsibility in value.items():
            if centroid not in updated:
                updated[centroid] = np.array(key) * responsibility
                denoms[centroid] = responsibility
            else:
                updated[centroid] += np.array(key) * responsibility
                denoms[centroid] += responsibility
    for centroid in updated:
        updated[centroid] /= denoms[centroid]
        updated[centroid] = updated[centroid].tolist()
    return updated
    # END_YOUR_ANSWER



def main(data, init_centroids):
    #######################################################
    # You do not need to change anything in this function #
    #######################################################
    beta = 3
    centroids = init_centroids
    old_centroids = None
    total_step = 7
    for step in range(total_step):
        # save old centroid
        old_centroids = centroids
        # new assignment
        soft_assignment_dict = update_soft_assignment(data, old_centroids, beta)
        # update centroids
        centroids = update_centroids(soft_assignment_dict)
        # plot centroid
        fig = plot_2d_soft(soft_assignment_dict, centroids)
        plt.title(f"step{step}")
        fig.savefig(os.path.join("results", "2D_soft", f"step{step}.png"))
        plt.clf()
    print(f"{total_step} iterations were completed.")
    return centroids


if __name__ == '__main__':
    data, label = read_data("data/data_2d.csv")
    init_c = load_centroids("data/2d_init_centroids.csv")
    final_c = main(data, init_c)
    write_centroids_tofile("2d_final_centroids_with_soft_kmeans.csv", final_c)
